app/requirements.txt
# SOCAR Hackathon LLM Endpoint Dependencies
# Optimized for production deployment

# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3

# Azure OpenAI client
openai==1.54.0

# Vector database
pinecone-client==5.0.0

# Embeddings
sentence-transformers==3.3.1
torch==2.1.2

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Optional: monitoring and logging
prometheus-fastapi-instrumentator==7.0.0
